requirements.txt
asgiref==3.2.10
attrs==19.3.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
constantly==15.1.0
cryptography==3.0
cssselect==1.1.0
Django==3.1
django-cors-headers==3.4.0
django-redis==4.12.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.2
jmespath==0.10.0
lxml==4.5.2
mysqlclient==2.0.1
parsel==1.6.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL @ file:///D:/fiction_website/source_code/pyOpenSSL-19.1.0-py2.py3-none-any.whl
pytz==2020.1
pywin32==228
queuelib==1.5.0
redis==3.5.3
requests==2.24.0
Scrapy==2.3.0
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
urllib3==1.25.10
w3lib==1.22.0
zope.interface==5.1.0
